Rename the source files for ChannelArgsEndpointConfig
Previous naming was non-descript
