+++ gRPC-C++.podspec
@@ -580,7 +580,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/debug/stats.h',
                       'src/core/lib/debug/stats_data.h',
                       'src/core/lib/debug/trace.h',
-                      'src/core/lib/event_engine/endpoint_config_internal.h',
+                      'src/core/lib/event_engine/channel_args_endpoint_config.h',
                       'src/core/lib/event_engine/sockaddr.h',
                       'src/core/lib/gpr/alloc.h',
                       'src/core/lib/gpr/env.h',
@@ -1273,7 +1273,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/debug/stats.h',
                               'src/core/lib/debug/stats_data.h',
                               'src/core/lib/debug/trace.h',
-                              'src/core/lib/event_engine/endpoint_config_internal.h',
+                              'src/core/lib/event_engine/channel_args_endpoint_config.h',
                               'src/core/lib/event_engine/sockaddr.h',
                               'src/core/lib/gpr/alloc.h',
                               'src/core/lib/gpr/env.h',
