+++ gRPC-Core.podspec
@@ -867,8 +867,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/debug/stats_data.h',
                       'src/core/lib/debug/trace.cc',
                       'src/core/lib/debug/trace.h',
+                      'src/core/lib/event_engine/channel_args_endpoint_config.h',
                       'src/core/lib/event_engine/endpoint_config.cc',
-                      'src/core/lib/event_engine/endpoint_config_internal.h',
                       'src/core/lib/event_engine/event_engine.cc',
                       'src/core/lib/event_engine/sockaddr.cc',
                       'src/core/lib/event_engine/sockaddr.h',
@@ -1789,7 +1789,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/debug/stats.h',
                               'src/core/lib/debug/stats_data.h',
                               'src/core/lib/debug/trace.h',
-                              'src/core/lib/event_engine/endpoint_config_internal.h',
+                              'src/core/lib/event_engine/channel_args_endpoint_config.h',
                               'src/core/lib/event_engine/sockaddr.h',
                               'src/core/lib/gpr/alloc.h',
                               'src/core/lib/gpr/env.h',
