+++ test/core/event_engine/endpoint_config_test.cc
@@ -20,7 +20,7 @@
 #include <grpc/grpc.h>
 
 #include "src/core/lib/channel/channel_args.h"
-#include "src/core/lib/event_engine/endpoint_config_internal.h"
+#include "src/core/lib/event_engine/channel_args_endpoint_config.h"
 #include "test/core/util/test_config.h"
 
 using ::grpc_event_engine::experimental::ChannelArgsEndpointConfig;
